Remove the signals module and the dependency to matplotlib.
We do not want this kind of functionality in a convex optimization
toolbox. The sole use of this module was for a tutorial to load an image,
move the code directly in the tutorial. The white circle image can even
be generated and do not need to be stored. We can then drop the
dependency to matplotlib which is only necessary to test the
reconstruction tutorial. It is thus installed by Travis, but not a
requirement.